+++ src/components/home/type.ts
